Extracted the emptyCard component into it's own module
